Rename to SoundPreferencesViewEventTarget
