+++ Telephone/PreferencesController.h
@@ -22,7 +22,7 @@
 #import "PreferencesControllerNotifications.h"
 #import "SoundIOPreferences.h"
 
-@protocol SoundPreferencesViewEventTarget;
+@class SoundPreferencesViewEventTarget;
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -43,7 +43,7 @@ enum {
 
 @property(nonatomic, readonly, weak) id<PreferencesControllerDelegate> delegate;
 @property(nonatomic, readonly) AKSIPUserAgent *userAgent;
-@property(nonatomic, readonly) id<SoundPreferencesViewEventTarget> soundPreferencesViewEventTarget;
+@property(nonatomic, readonly) SoundPreferencesViewEventTarget *soundPreferencesViewEventTarget;
 
 // General preferences view controller.
 @property(nonatomic, readonly) GeneralPreferencesViewController *generalPreferencesViewController;
@@ -67,7 +67,7 @@ enum {
 
 - (instancetype)initWithDelegate:(id<PreferencesControllerDelegate>)delegate
                        userAgent:(AKSIPUserAgent *)userAgent
- soundPreferencesViewEventTarget:(id<SoundPreferencesViewEventTarget>)soundPreferencesViewEventTarget;
+ soundPreferencesViewEventTarget:(SoundPreferencesViewEventTarget *)soundPreferencesViewEventTarget;
 
 // Changes window's content view.
 - (IBAction)changeView:(id)sender;
